Seperate the ust_info_scraper.js into 2 js files - ust_link.js: generate link,json - ust_info_scraper.js: generate the final.json
